code re-organization
